+++ src/components/header/pair-table.tsx
