+++ src/components/market-data/trade-history.tsx
