+++ src/components/terminal.tsx
@@ -2,10 +2,10 @@ import { getRouteApi, useNavigate } from "@tanstack/react-router";
 
 import { usePools } from "@/hooks/usePools";
 import { PoolContext } from "@/contexts/pool";
-import Header from "@/components/header";
+import Header from "@/components/header/header";
 import Chart from "@/components/chart";
 import Trade from "@/components/trade/trade";
-import MarketOverview from "@/components/market-overview";
+import MarketOverview from "@/components/market-data/market-data";
 import User from "@/components/user/user";
 import { Toaster } from "@/components/ui/toaster";
